Move `download_file()` from ollama.py to common.py
